Documentation, misc. release fixes
README updates:
- Rewriting README, added new example images
- Exclude GLID code from IntraPaint bundle
- API fix for latest webui-forge
- Rename A1111 WebUI config category to Remote Stable-Diffusion Config
- Added configurable backup path for loading libmypaint
- Added is_pyinstaller_bundle utility function

+++ IntraPaint.py
@@ -14,6 +14,7 @@ from PySide6.QtWidgets import QApplication, QSplashScreen
 
 from src.util.geometry_utils import get_scaled_placement
 from src.util.optional_import import check_import
+from src.util.pyinstaller import is_pyinstaller_bundle
 from src.util.shared_constants import TIMELAPSE_MODE_FLAG, PROJECT_DIR, LOG_DIR
 from src.util.arg_parser import build_arg_parser
 
@@ -101,28 +102,29 @@ for root, _, files in os.walk(f'{PROJECT_DIR}/resources/translations'):
 app.installTranslator(translator)
 
 # If relevant directories exist, update paths for GLID-3-XL dependencies:
-if not check_import('ldm'):
-    expected_ldm_path = f'{PROJECT_DIR}/latent-diffusion'
-    if os.path.exists(expected_ldm_path):
-        sys.path.append(expected_ldm_path)
+if not is_pyinstaller_bundle():
+    if not check_import('ldm'):
+        expected_ldm_path = f'{PROJECT_DIR}/latent-diffusion'
+        if os.path.exists(expected_ldm_path):
+            sys.path.append(expected_ldm_path)
 
-        # Newer versions of pytorch-lightning changed the location of one needed dependency, but latent-diffusion was
-        # never updated. This only requires a single minor update, so make that change here if necessary:
-        updated_file_path = f'{expected_ldm_path}/ldm/models/diffusion/ddpm.py'
-        with open(updated_file_path, 'r+') as module_file:
-            lines = module_file.readlines()
-            for i, line in enumerate(lines):
-                if 'from pytorch_lightning.utilities.distributed import rank_zero_only' in line:
-                    lines[i] = 'from pytorch_lightning.utilities.rank_zero import rank_zero_only'
-                    module_file.seek(0)
-                    module_file.writelines(lines)
-                    module_file.truncate()
-                    break
+            # Newer versions of pytorch-lightning changed the location of one needed dependency, but latent-diffusion was
+            # never updated. This only requires a single minor update, so make that change here if necessary:
+            updated_file_path = f'{expected_ldm_path}/ldm/models/diffusion/ddpm.py'
+            with open(updated_file_path, 'r+') as module_file:
+                lines = module_file.readlines()
+                for i, line in enumerate(lines):
+                    if 'from pytorch_lightning.utilities.distributed import rank_zero_only' in line:
+                        lines[i] = 'from pytorch_lightning.utilities.rank_zero import rank_zero_only'
+                        module_file.seek(0)
+                        module_file.writelines(lines)
+                        module_file.truncate()
+                        break
 
-if not check_import('taming'):
-    expected_taming_path = f'{PROJECT_DIR}/taming-transformers'
-    if os.path.exists(expected_taming_path):
-        sys.path.append(expected_taming_path)
+    if not check_import('taming'):
+        expected_taming_path = f'{PROJECT_DIR}/taming-transformers'
+        if os.path.exists(expected_taming_path):
+            sys.path.append(expected_taming_path)
 
 # These imports need to be delayed until after logging setup, translation, and import path tweaks:
 # noinspection PyPep8
